refactor(galaxy-toolshed): version, split routes for repository, requirement, and tool
